Ollama-Prototype/requirements.txt
ollama
python-dotenv
pydantic
tqdm
scikit-learn
